Roll convenient.rs into statement.rs.
